+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* [`useValidateAsSetter`] Added a hook that supports using `validate` from `useFrom` with
+    `useSettersAsRefEventHandler`, and `useSettersAsEventHandler`
 
 ## [0.0.10] - 2019-12-14
 ### Added